'logic-added-to-handle-client-state-management-for-updated-entry-together-with-analysis'
