+++ app/api/journal/[id]/route.ts
@@ -1,6 +1,7 @@
 import { analyze } from "@/utils/ai";
 import { getUserByClerkId } from "@/utils/auth";
 import { prisma } from "@/utils/db";
+import { revalidatePath } from "next/cache";
 import { NextResponse } from "next/server";
 
 export const PATCH = async (request: Request, { params }) => {
@@ -18,7 +19,7 @@ export const PATCH = async (request: Request, { params }) => {
     },
   });
 
-  const analyis = await analyze(updatedEntry.content);
+  const analysis = await analyze(updatedEntry.content);
 
   const updated = await prisma.analysis.upsert({
     where: {
@@ -26,12 +27,10 @@ export const PATCH = async (request: Request, { params }) => {
     },
     create: {
       entryId: updatedEntry.id,
-      ...analyis,
+      ...analysis,
     },
-    update: analyis,
+    update: analysis,
   });
 
-  console.log(updated);
-
-  return NextResponse.json({ data: updatedEntry });
+  return NextResponse.json({ data: { ...updatedEntry, analysis: updated } });
 };
